aio_server/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
aiofiles==23.2.1
SQLAlchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
httpx==0.25.1 